+++ proj-android/PowerAuthLibrary/src/main/java/io/getlime/security/powerauth/networking/response/IActivationRenameListener.java
@@ -1,0 +1,42 @@
+/*
+ * Copyright 2026 Wultra s.r.o.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.getlime.security.powerauth.networking.response;
+
+import androidx.annotation.MainThread;
+import androidx.annotation.NonNull;
+
+/**
+ * Listener for activation rename.
+ */
+public interface IActivationRenameListener {
+
+    /**
+     * Called when activation is successfully renamed.
+     *
+     * @param activationName Server-confirmed activation name.
+     */
+    @MainThread
+    void onActivationRenameSucceed(@NonNull String activationName);
+
+    /**
+     * Called when activation rename fails.
+     *
+     * @param t error that occurred during the activation rename.
+     */
+    @MainThread
+    void onActivationRenameFailed(@NonNull Throwable t);
+}